codes/requirements.txt
h5py==3.7.0
numpy==1.21.1
opencv_python==4.6.0.66
Pillow==9.3.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
